Moved google-drive-query-encoder back into google-realtime.service.ts
