--- src/client/app/service/google-realtime/google-drive-query-encoder.ts
@@ -1,11 +1,0 @@
-import {QueryEncoder} from '@angular/http';
-
-export class GoogleDriveQueryEncoder extends QueryEncoder {
-  encodeKey(key: string) {
-    return encodeURIComponent(key);
-  }
-
-  encodeValue(value: string) {
-    return encodeURIComponent(value);
-  }
-}
